requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.13
chardet==3.0.4
coverage==4.5.1
Django
djangorestframework==3.8.2
eradicate==0.2.1
factory-boy==2.11.1
Faker==0.9.1
get==1.0.3
html5lib==1.0.1
idna==2.7
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.7.1
post==1.0.2
public==1.0.3
py==1.6.0
pycodestyle==2.4.0
pydocstyle==2.1.1
pyflakes==2.0.0
pylama==7.5.5
pytest==3.8.1
pytest-django==3.4.3
python-dateutil==2.7.3
pytz==2018.5
query-string==1.0.2
requests==2.19.1
six==1.11.0
snowballstemmer==1.2.1
text-unidecode==1.2
toml==0.9.6
tox==3.4.0
urllib3==1.23
virtualenv==16.0.0
webencodings==0.5.1
